+++ module-apps/application-desktop/windows/LockedInfoWindow.hpp
@@ -6,16 +6,16 @@
 
 namespace gui
 {
-
     class LockedInfoWindow : public AppWindow
     {
         gui::Image *lockImage = nullptr;
         gui::Text *infoText   = nullptr;
-        // method hides or show widgets and sets bars according to provided state
+
         void setVisibleState();
         void invalidate() noexcept;
+
       public:
-        LockedInfoWindow(app::Application *app);
+        explicit LockedInfoWindow(app::Application *app);
         void onBeforeShow(ShowMode mode, SwitchData *data) override;
         bool onInput(const InputEvent &inputEvent) override;
 
@@ -23,5 +23,4 @@ namespace gui
         void buildInterface() override;
         void destroyInterface() override;
     };
-
 } /* namespace gui */