+++ module-gui/gui/widgets/RichTextParser.cpp
@@ -236,10 +236,10 @@ namespace text
         }
     };
 
-    class ShortenedTextNodes
+    class ShortTextNodes
     {
-        using singleAttributedNode = std::map<std::string, std::pair<std::string, std::string>> const;
-        static singleAttributedNode nodes;
+        using SingleAttributedNode = std::map<std::string, std::pair<std::string, std::string>>;
+        static const SingleAttributedNode nodes;
 
       public:
         enum class AttributeContent
@@ -248,27 +248,27 @@ namespace text
             Value
         };
 
-        [[nodiscard]] static auto is(const char *nodeName)
+        [[nodiscard]] static auto is(const char *nodeName) -> bool
         {
-            return nodes.count(std::string(nodeName)) > 0;
+            return static_cast<bool>(nodes.count(std::string(nodeName)));
         }
 
-        [[nodiscard]] static auto get(const char *nodeName, AttributeContent content)
+        [[nodiscard]] static auto get(const char *nodeName, AttributeContent content) -> std::optional<std::string>
         {
             try {
-                if (content == AttributeContent::Name) {
-                    return nodes.at(nodeName).first;
-                }
-                return nodes.at(nodeName).second;
+                auto attribute = nodes.at(nodeName);
+                return std::optional<std::string>(content == AttributeContent::Name ? attribute.first
+                                                                                    : attribute.second);
             }
             catch (std::out_of_range &) {
-                return std::string{};
+                LOG_ERROR("ShortTextNode: %s not found", nodeName);
+                return {};
             }
         }
     };
 
-    ShortenedTextNodes::singleAttributedNode ShortenedTextNodes::nodes = {
-        {gui::text::shortened_bold, {gui::text::weight, gui::text::bold}}};
+    const ShortTextNodes::SingleAttributedNode ShortTextNodes::nodes = {
+        {gui::text::short_bold, {gui::text::weight, gui::text::bold}}};
 }; // namespace text
 
 struct walker : pugi::xml_tree_walker
@@ -309,9 +309,9 @@ struct walker : pugi::xml_tree_walker
         return std::string(node.name()) == gui::text::node_br || std::string(node.name()) == gui::text::node_p;
     }
 
-    auto is_shortened_text_node(pugi::xml_node &node) const
+    auto is_short_text_node(pugi::xml_node &node) const
     {
-        return text::ShortenedTextNodes::is(node.name());
+        return text::ShortTextNodes::is(node.name());
     }
 
     auto push_text_node(pugi::xml_node &node)
@@ -326,18 +326,19 @@ struct walker : pugi::xml_tree_walker
         log_parser("Attr loaded: %s", style_stack.back().str().c_str());
     }
 
-    auto push_text_shortened_node(pugi::xml_node &node)
+    auto push_short_text_node(pugi::xml_node &node)
     {
         log_parser("shortened text node name: %s", node.name());
         auto local_style = style_stack.back();
         auto &decor      = text::NodeDecor::get();
 
-        using stNodes   = text::ShortenedTextNodes;
-        using stContent = stNodes::AttributeContent;
-        decor.stack_visit(
-            local_style, stNodes::get(node.name(), stContent::Name), stNodes::get(node.name(), stContent::Value));
-        style_stack.push_back(local_style);
-        log_parser("Attr loaded: %s", style_stack.back().str().c_str());
+        auto attrName  = text::ShortTextNodes::get(node.name(), text::ShortTextNodes::AttributeContent::Name);
+        auto attrValue = text::ShortTextNodes::get(node.name(), text::ShortTextNodes::AttributeContent::Value);
+        if (attrName.has_value() && attrValue.has_value()) {
+            decor.stack_visit(local_style, attrName.value(), attrValue.value());
+            style_stack.push_back(local_style);
+            log_parser("Attr loaded: %s", style_stack.back().str().c_str());
+        }
     }
 
     auto push_newline_node(pugi::xml_node &)
@@ -360,8 +361,8 @@ struct walker : pugi::xml_tree_walker
                 push_text_node(node);
                 return true;
             }
-            if (is_shortened_text_node(node)) {
-                push_text_shortened_node(node);
+            if (is_short_text_node(node)) {
+                push_short_text_node(node);
                 return true;
             }
             if (is_newline_node(node)) {
@@ -394,7 +395,7 @@ struct walker : pugi::xml_tree_walker
         log_node(node, Action::Exit);
 
         if (node.type() == pugi::xml_node_type::node_element) {
-            if (std::string(node.name()) == gui::text::node_text || is_shortened_text_node(node)) {
+            if (std::string(node.name()) == gui::text::node_text || is_short_text_node(node)) {
                 pop_text_node(node);
                 return true;
             }